refactor analyze

fix lint


fix lint
